requirements.txt
astroid==2.8.0
attrs==21.2.0
dnspython==1.16.0
flake8==3.9.2
googleapis-common-protos==1.53.0
grpcio==1.41.0
grpcio-status==1.41.0
grpcio-tools==1.41.0
iniconfig==1.1.1
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
packaging==21.0
platformdirs==2.4.0
pluggy==1.0.0
protobuf==3.18.0
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
pylint==2.11.1
pymongo==3.12.0
pyparsing==2.4.7
pytest==6.2.5
PyYAML==5.4.1
six==1.16.0
toml==0.10.2
typing-extensions==3.10.0.2
wrapt==1.12.1
